requirements.txt
# Core dependencies
pydantic>=2.6.0
pydantic-settings>=2.1.0
pydantic_core>=2.16.0
typing_extensions>=4.9.0

# PDF processing
pdfplumber>=0.11.0
pdfminer.six>=20251230
pypdfium2>=5.3.0

# Word document processing
python-docx>=1.1.0
docxtpl>=0.16.7

# Excel processing
openpyxl>=3.1.0
xlsxwriter>=3.2.0
pandas>=2.2.0

# PowerPoint processing
python-pptx>=1.0.0

# Jinja2 templating
jinja2>=3.1.0
MarkupSafe>=3.0.0

# Environment variable management
python-dotenv>=1.0.0

# Web framework (if needed for API endpoints)
fastapi>=0.115.0
uvicorn>=0.32.0
starlette>=0.40.0
python-multipart>=0.0.20

# HTTP client
httpx>=0.28.0
httpcore>=1.0.0
requests>=2.31.0

# YAML parsing
PyYAML>=6.0.0

# CLI tools
typer>=0.12.0
click>=8.1.0
rich>=13.7.0
colorlog>=6.10.0

# JSON utilities
ujson>=5.11.0

# File handling
aiofiles>=25.0.0
fsspec>=2026.1.0

# OCR and image processing (if needed)
paddleocr>=3.3.0
paddlepaddle>=3.2.0
opencv-contrib-python>=4.10.0
pillow>=12.0.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.9.0
pytz>=2025.0
